requirements.txt
Flask
torch
torchvision
Pillow
numpy
matplotlib
seaborn
scikit-learn
